refactor(gym/executor): split tests to sibling tests_executor_metrics.rs (#1266)
Inline test module in executor.rs is moved into a NEW sibling file
tests_executor_metrics.rs (rather than appending to existing
tests_executor.rs which is already 629 LOC and would otherwise grow
past 400).

- src/gym/executor.rs: 426 → 358 LOC (under 400-line rule)
- src/gym/tests_executor_metrics.rs: new (~67 LOC)
- src/gym/mod.rs: register mod tests_executor_metrics
- super::* → super::executor::* (sibling level)

Pure refactor: zero logic changes, all 64 gym::tests_executor* tests pass.

Closes 3/70 of #1266 (this loop iteration), 11 PRs cumulative since #1285.
